requirements.txt
# 环境变量 配置信息
python-dotenv
yaml

# mongodb数据库
motor

# 实验追踪
mlflow

# 数据处理
pandas
numpy
scikit-learn

# 后端
fastapi
uvicorn

# 便捷性
python-box

# 测试
pytest
httpx